BUG: Fix error "Remove-Alias : The term 'Remove-Alias' is not recognized as the name of a cmdlet, function, script file, or operable program. Check the spelling of the name, or if a path was included, verify that the path is correct and try again." when starting Windows PowerShell.

+++ Documents/WindowsPowerShell/profile.ps1
@@ -82,7 +82,9 @@ if (Test-Path($ChocolateyProfile)) {
 }
 
 # Rename default aliases.
-Remove-Alias h
+if ($PSVersionTable.PSVersion -ge [Version]"6.0") {
+	Remove-Alias h
+}
 Set-Alias hi Get-History
